requirements.txt
pypdf>=3.0.0
langchain>=0.1.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chainlit>=0.7.0
huggingface-hub>=0.19.0
langchain-community>=0.0.10
python-dotenv>=1.0.0